requirements.txt
numpy
openai-whisper
ollama
ffmpeg-python
